% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_binding_isotherm.R
\name{augment_shared_isotherms}
\alias{augment_shared_isotherms}
\title{Augmenting function for shared isotherm fits}
\usage{
augment_shared_isotherms(x, newdata, INDEX = NULL)
}
\arguments{
\item{x}{A fitted model returned from \code{\link{fit_binding_isotherm}}}

\item{newdata}{The new data to fit; must use the same column names that
were used to create the model.}

\item{INDEX}{A column variable that shall be used to split the data into at
most two sets, for which both will share higher order K. You will get other
estimates as .x and .y fits.}
}
\description{
Augmenting function for shared isotherm fits
}
