% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_binding_isotherm.R
\name{gpf_fraction_bound}
\alias{gpf_fraction_bound}
\title{Compute the bound fraction from the grand partition function}
\usage{
gpf_fraction_bound(x, binding_constants, type = "macro")
}
\arguments{
\item{x}{A vector of ligand concentrations.}

\item{binding_constants}{The values of the association equilibrium constant
given as a (named) vector. See Details.}

\item{type}{Either \code{"micro"} or \code{"macro"} (default).}
}
\description{
Compute the bound fraction from the grand partition function
}
\details{
If \code{type == "macro"}, the binding isotherm is computed for a macroscopic
process using \code{\link{gpf_macro}}. The entries are assumed to be in ascending
order, i.e., \code{binding_constants[2]} gives the apparent thermodynamic binding
constant for associating the second ligand.

If \code{type == "micro"} the vector must contain the (named) microsocopic
binding constants compatible with \code{\link{gpf_micro}}. Note that in this
case, the entries are assumed to be in combinatorically increasing order,
i.e., "a", "b", "c", "ab", "ac", "bc", "abc", for a third degree polynom.
The entries can be named, but the names are not used for the assignment due
to the complexity of the problem.

In either case, the degree is determined from the length of the binding constants
vector.
}
