% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_binding_isotherm.R
\name{gpf_fraction_micro_macro}
\alias{gpf_fraction_micro_macro}
\title{Quick overview plots to demonstrate harmonic mean}
\usage{
gpf_fraction_micro_macro(a, b, c = 0, x = 10^seq(-6, 2, length.out = 100))
}
\description{
Quick overview plots to demonstrate harmonic mean
}
