% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_binding_isotherm.R
\name{gpf_macro}
\alias{gpf_macro}
\title{Get macroscopic binding polynomial}
\usage{
gpf_macro(
  degree = 2,
  params = as.character(seq(degree)),
  xname = "x",
  kname = "K_"
)
}
\arguments{
\item{degree}{The degree of the polynom; overwritten by \code{params}.}

\item{params}{A character vector of length i to model i binding states; the
associated equilibrium constants take these names.}

\item{xname}{The name of the variable representing the bound/titrated ligand.}

\item{kname}{The common variable name for the equilibrium constants.}
}
\value{
An expression.
}
\description{
Returns an expression that represents the macroscopic (thermodynamic) binding
polynomial (the grand canonical partition function) for an idealized system
that binds up to i ligands.
}
