% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_binding_isotherm.R
\name{gpf_micro}
\alias{gpf_micro}
\title{Get microscopic binding polynomial}
\usage{
gpf_micro(degree = 2, params = letters[seq(degree)], xname = "x", kname = "k_")
}
\arguments{
\item{degree}{The degree of the polynom; overwritten by \code{params}.}

\item{params}{A character vector of length i to model i binding sites; the
associated equilibrium constants take these indices.}

\item{xname}{The name of the variable representing the bound/titrated ligand.}

\item{kname}{The common variable name for the equilibrium constants.}
}
\value{
An expression.
}
\description{
Returns an expression that represents the microscopic binding
polynomial for an idealized system with i (different) binding sites.
}
\details{
As compared to \code{\link{gpf_macro}}, the returned expression has
more variables than constructed from \code{degree} or passed via \code{params}.
}
