% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqtl_fun.R
\name{iqtl_import_all}
\alias{iqtl_import_all}
\title{Import IQTL files from a list}
\usage{
iqtl_import_all(l, path = NULL, ...)
}
\arguments{
\item{l}{A list with descriptors. See examples.}

\item{path}{A path in which all \code{file}s are residing. If \code{NULL},
\code{file}s should be fully expanded paths or refer to the current working
directory.}

\item{...}{
  Arguments passed on to \code{\link[=iqtl_meta]{iqtl_meta}}
  \describe{
    \item{\code{meta_data}}{List to construct a metadata data frame from.}
    \item{\code{exclude}}{A character vector with \code{lane_id}s to remove from the result.}
    \item{\code{file}}{Path to the file.}
  }}
}
\description{
Given a \code{list} or \code{data.frame} with an entry/column \code{file},
this function feeds its arguments to \code{\link{iqtl_meta}}.
}
\details{
Entries named \code{.$file} or \code{.$exclude} are passed as separate arguments
to \code{\link{iqtl_meta}}. All other entries of \code{l} are passed to
\code{iqtl_meta} as if they were given as \code{meta_data} list.

Empty entries are dropped without a warning. Missing files are dropped with
a warning.
}
\examples{

}
