% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqtl_fun.R
\name{iqtl_meta}
\alias{iqtl_meta}
\title{Associate metadata with an IQTL file}
\usage{
iqtl_meta(file, meta_data = list(conc = NA), exclude = NULL, ...)
}
\arguments{
\item{file}{Path to the file.}

\item{meta_data}{List to construct a metadata data frame from.}

\item{exclude}{A character vector with \code{lane_id}s to remove from the result.}

\item{...}{Other arguments passed to \code{iqtl_read}.}
}
\value{
A data frame of class \code{\link[data.table:data.table]{data.table}}.
}
\description{
Add metadata along an imported IQTL data file.
}
\details{
This function calls \code{iqtl_read} with its default arguments for
\code{col_group} ("lane"), \code{.id} ("id") and \code{.sep} ("_").

If \code{meta_data} contains an entry named \code{"lane_id"}, its entries are
used to join the imported data with the metadata annotation. If no such entry
exists, the data is assumed to be parallel to the result of \code{iqtl_read}.
(More specifically, to the \code{levels} of \code{.$lane_id}.)

Internally, \code{meta_data} is converted to a \code{data.frame}.
}
