% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqtl_fun.R
\name{iqtl_read}
\alias{iqtl_read}
\title{Import an IQTL file}
\usage{
iqtl_read(
  file,
  row_name = "band",
  col_group = "lane",
  col_repeat = c(quantify = "vol", "vol+bkg", "CAL", "MW", "Rf"),
  reverse_rows = TRUE,
  .id = "id",
  .sep = "_",
  gel_name = paste0("gel", .sep, "[[:alnum:]_-]+")
)
}
\arguments{
\item{file}{Path to the file.}

\item{row_name}{Name to give to rows.}

\item{col_group}{Name to give to column groups, i.e., repeating units of \code{col_repeat}.}

\item{col_repeat}{Column variables that repeat along \code{col_group}.}

\item{reverse_rows}{If \code{TRUE}, the "\code{row_name}" identifiers are counted
from bottom to top; else from top to bottom.}

\item{.id}{General identifier added to \code{row_name} and \code{col_group}.}

\item{.sep}{General separator.}

\item{gel_name}{A \link[base:regex]{regular expression} to extract an identifier
from the file name.}
}
\value{
A data frame of class \code{\link[data.table:data.table]{data.table}}.
}
\description{
Reads a file exported with ImageQuant TL and performs basic actions to tidy
up the format.
}
\examples{
assaydata <- system.file("extdata", "gel_01.txt", package = "summerrband")

iqtl_read(assaydata)

}
