% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqtl_fun.R
\name{iqtl_view}
\alias{iqtl_view}
\title{Import and view an IQTL file}
\usage{
iqtl_view(file, selected_bands = character(), ..., on_y_axis = "vol_frac")
}
\arguments{
\item{file}{Path to the file.}

\item{selected_bands}{Character vector to imit plotting to specific bands,
e.g. \code{"band_0"}.}

\item{...}{Other arguments passed to \code{iqtl_read}.}

\item{on_y_axis}{Which variable to draw on the y axis.}
}
\value{
A data frame of class \code{\link[data.table:data.table]{data.table}}.
}
\description{
Reads a file exported with ImageQuant TL and performs basic actions to tidy
up the format using \code{\link{iqtl_read}} and draws an overview plot.
}
\details{
This function calls \code{iqtl_read} with its default arguments for \code{row_name}
("band"), \code{col_group} ("lane"), \code{.id} ("id") and \code{.sep} ("_").
It returns the same object as a default call to \code{iqtl_read}.
As a side-effect, a plot is printed to the current device.
}
