% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summerrband.R
\docType{package}
\name{summerrband-package}
\alias{summerrband-package}
\title{summerrband}
\description{
Analyze gel-shift assays.

\subsection{Importing IQTL Data}{

\link{iqtl_read}, \link{iqtl_meta}, \link{iqtl_import_all}

\link{iqtl_view} to quickly view an IQTL file

}

\subsection{Workflows}{

See \link{use_template} for workflows involving, e.g., fitting of Kd values (\link{fit_Kd}).

}
}
\details{
summerrband-package
}
\author{
Benjamin Buchmuller \email{benjamin.buchmuller@tu-dortmund.de}
}
