% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summerrband.R
\name{use_template}
\alias{use_template}
\title{Use a template for a IQTL/EMSA script}
\usage{
use_template(version = "A01")
}
\arguments{
\item{version}{A template version identifier.}
}
\description{
Use a template for a IQTL/EMSA script
}
\details{
\describe{

\item{A01}{EMSA with selectivity measurements}

\item{A02}{EMSA with affinity measurements and Kd fitting}

}
}
