% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summerrband.R
\docType{package}
\name{summerrband-package}
\alias{summerrband-package}
\title{summerrband}
\description{
Analyze gel-shift assays.
}
\details{
summerrband-package

This package provides functions to evaluate gel-shift assays, e.g. to determine
binding affinity.
}
\author{
Benjamin Buchmuller \email{benjamin.buchmuller@tu-dortmund.de}
}
